fix OUT endpoints
was using only lower 8 bits of configuration registers

+++ hardware/arduino/sam/libraries/HID/HID.cpp
@@ -144,7 +144,7 @@ bool HID_Setup(USBSetup& setup, uint8_t i)
 
 HID_::HID_(void)
 {
-	static uint8_t endpointType[1];
+	static uint32_t endpointType[1];
 
 	endpointType[0] = EP_TYPE_INTERRUPT_IN;
